Leverage nlbuild-autotools-1.5.0 and allow on-demand remote synchronization of nlunit-test from an external repository when '--with-nlunit_test=internal' is asserted.
NLUNIT_TEST_SUBDIRS is not a permanent part of DIST_SUBDIRS since we do not ever want to include it in a distribution archive; however, when it's been pulled as a 'repo' module, we do want to remove it on invocation of the 'distclean' target. Consequently, we conditionally include it in DIST_SUBDIRS on invocation of 'distclean-recursive'.